+++ src/main/java/com/amazonaws/services/neptune/cli/PropertyGraphSerializationModule.java
@@ -14,19 +14,17 @@ package com.amazonaws.services.neptune.cli;
 
 import com.amazonaws.services.neptune.propertygraph.NeptuneGremlinClient;
 import com.amazonaws.services.neptune.propertygraph.io.SerializationConfig;
-import com.amazonaws.services.neptune.propertygraph.schema.TokensOnly;
 import com.github.rvesse.airline.annotations.Option;
 import com.github.rvesse.airline.annotations.restrictions.AllowedEnumValues;
-import com.github.rvesse.airline.annotations.restrictions.AllowedValues;
 import com.github.rvesse.airline.annotations.restrictions.Once;
-import org.apache.tinkerpop.gremlin.driver.ser.Serializers;
+import org.apache.tinkerpop.gremlin.util.ser.Serializers;
 
 public class PropertyGraphSerializationModule {
 
     @Option(name = {"--serializer"}, description = "Message serializer – (optional, default 'GRAPHBINARY_V1D0').")
     @AllowedEnumValues(Serializers.class)
     @Once
-    private String serializer = Serializers.GRAPHBINARY_V1D0.name();
+    private String serializer = Serializers.GRAPHBINARY_V1.name();
 
     @Option(name = {"--janus"}, description = "Use JanusGraph serializer.")
     @Once